Cleanup of Metronome and SimpleMetronome. Added documentation and moved classes to own package.
