--- stabilizer/src/main/java/com/hazelcast/stabilizer/worker/Metronome.java
@@ -1,5 +1,0 @@
-package com.hazelcast.stabilizer.worker;
-
-public interface Metronome {
-    void waitForNext();
-}
