+++ stabilizer/src/main/java/com/hazelcast/stabilizer/worker/metronome/Metronome.java
@@ -1,0 +1,12 @@
+package com.hazelcast.stabilizer.worker.metronome;
+
+/**
+ * Used to clock a running task or worker with a defined interval.
+ */
+public interface Metronome {
+
+    /**
+     * Waits for the defined interval.
+     */
+    void waitForNext();
+}
