+++ stabilizer/src/test/java/com/hazelcast/stabilizer/worker/metronome/SimpleMetronomeTest.java
@@ -1,4 +1,4 @@
-package com.hazelcast.stabilizer.worker;
+package com.hazelcast.stabilizer.worker.metronome;
 
 import org.junit.Test;
