+++ tests/src/main/java/com/hazelcast/stabilizer/tests/map/SqlPredicateTest.java
@@ -17,8 +17,8 @@ import com.hazelcast.stabilizer.test.annotations.Setup;
 import com.hazelcast.stabilizer.test.annotations.Teardown;
 import com.hazelcast.stabilizer.test.annotations.Warmup;
 import com.hazelcast.stabilizer.test.utils.ThreadSpawner;
-import com.hazelcast.stabilizer.worker.Metronome;
-import com.hazelcast.stabilizer.worker.SimpleMetronome;
+import com.hazelcast.stabilizer.worker.metronome.Metronome;
+import com.hazelcast.stabilizer.worker.metronome.SimpleMetronome;
 
 import java.io.IOException;
 import java.util.Random;
